[AIRFLOW-6721] Move PrestoCheckOperator test
Move the PrestoCheckOperator test from the Hive operators module
into a new Presto operators test module while still using the
shared Hive test environment.
